Proximity detection for InteractObjects
First I added the interactObject script to the interact object prefab.

Then I added the interact object prefab to the scene

Then I made it so the player can actually move

Finally I made it so that if the player is within .5 units of the interact object a console message is displayed.